Update neoforge 26.1.2
完成 26.1.2 迁移：编译通过并产出 thetitansneo-1.0.8.jar

- 结构生成：Structure.generate(Holder<Structure>, ResourceKey<Level>, ...)、
  ChunkGenerator.createStructures(..., ResourceKey<Level>)；EndSpike -> EndSpikeFeature.EndSpike，
  SpikeConfiguration -> EndSpikeConfiguration
- 渲染收尾：getTextureLocation 改 public、LightningBolt 用渲染器缓存实体、
  RenderTypes.entityCutout/entityTranslucent 替代已删的 entityCutoutNoCull/entitySmoothCutout、
  InventoryMenu.BLOCK_ATLAS -> TextureAtlas.LOCATION_BLOCKS
- 特殊物品渲染器接 SpecialModelRenderer<ItemStack>
- FMLEnvironment.dist -> getDist()；Registries.TRIGGER_TYPES -> TRIGGER_TYPE
- mixin compatibilityLevel JAVA_21 -> JAVA_25（JDK 25）
- build.gradle/gradle.properties 补 neo_version_range 模板变量

+++ src/main/java/net/byAqua3/thetitansneo/entity/titan/EntityWitherzilla.java
@@ -789,7 +789,7 @@ public class EntityWitherzilla extends EntityTitan implements RangedAttackMob, I
 
 		if (!this.level().isClientSide()) {
 			ServerLevel serverLevel = (ServerLevel) this.level();
-			serverLevel.getServer().clockManager().addTicks(serverLevel.getServer().registryAccess().getOrThrow(net.minecraft.world.clock.WorldClocks.OVERWORLD), 18000L - serverLevel.getOverworldClockTime() % 24000L);
+			serverLevel.getServer().clockManager().addTicks(serverLevel.getServer().registryAccess().getOrThrow(net.minecraft.world.clock.WorldClocks.OVERWORLD), (int) (18000L - serverLevel.getOverworldClockTime() % 24000L));
 		}
 
 		if (this.getTarget() != null && !this.getTarget().isAlive()) {
